requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
tweepy==4.1.0
urllib3==1.26.7
